Create database.json
This is where the magic happens.
This file runs through all posts in the site.posts list and outputs a json-file with data on what word is used in what post.
The first node of the json file is called "postlist" and is a list of posts. Each list item is an object with a "title" and "url" key which are the title and relative url to the post, plus an optional "thumb" and "image" key which are urls to a thumbnail image and header image. These are filled in if the post's front matter specifies these keys.
The second node is called "searchdata" and contains an object of keys. Each key is a word that occurs in the posts, and the value of the key is a list of numbers representing the post where the word is found. Each of these numbers point to a position in the "postlist" list.
With this information the front end can quickly find what posts contain the queried word, and access the basic information about the post.
This file can probably be modified to include data from site.pages or other data you may want to search for.